+++ azext_iot/operations/hub.py
@@ -26,7 +26,7 @@ from azext_iot.common.shared import (
     KeyType,
     SettleType,
     RenewKeyType,
-    IoTHubStateType
+    IoTHubStateType,
 )
 from azext_iot.iothub.providers.discovery import IotHubDiscovery
 from azext_iot.common.utility import (
@@ -38,7 +38,7 @@ from azext_iot.common.utility import (
     init_monitoring,
     process_json_arg,
     ensure_iothub_sdk_min_version,
-    generate_key
+    generate_key,
 )
 from azext_iot._factory import SdkResolver, CloudError
 from azext_iot.operations.generic import _execute_query, _process_top
@@ -51,12 +51,21 @@ logger = get_logger(__name__)
 
 
 def iot_query(
-    cmd, query_command, hub_name=None, top=None, resource_group_name=None, login=None
+    cmd,
+    query_command,
+    hub_name=None,
+    top=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     top = _process_top(top)
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -74,11 +83,19 @@ def iot_query(
 
 
 def iot_device_show(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     return _iot_device_show(target, device_id)
 
@@ -104,13 +121,23 @@ def iot_device_list(
     edge_enabled=False,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     query = (
         "select * from devices where capabilities.iotEdge = true"
         if edge_enabled
         else "select * from devices"
     )
-    result = iot_query(cmd, query, hub_name, top, resource_group_name, login=login)
+    result = iot_query(
+        cmd=cmd,
+        query_command=query,
+        hub_name=hub_name,
+        top=top,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type_dataplane=auth_type_dataplane,
+    )
+
     if not result:
         logger.info('No registered devices found on hub "%s".', hub_name)
     return result
@@ -128,35 +155,19 @@ def iot_device_create(
     status_reason=None,
     valid_days=None,
     output_dir=None,
-    set_parent_id=None,
-    add_children=None,
-    force=False,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
-
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
-
-    if add_children:
-        if not edge_enabled:
-            raise CLIError(
-                'The device "{}" should be edge device in order to add children.'.format(device_id)
-            )
-
-        for child_device_id in add_children.split(","):
-            child_device = _iot_device_show(target, child_device_id.strip())
-            _validate_parent_child_relation(child_device, force)
-
-    deviceScope = None
-    if set_parent_id:
-        edge_device = _iot_device_show(target, set_parent_id)
-        _validate_edge_device(edge_device)
-        deviceScope = edge_device["deviceScope"]
 
     if any([valid_days, output_dir]):
         valid_days = 365 if not valid_days else int(valid_days)
@@ -179,18 +190,12 @@ def iot_device_create(
             secondary_thumbprint,
             status,
             status_reason,
-            deviceScope
         )
         output = service_sdk.devices.create_or_update_identity(
             id=device_id, device=device
         )
     except CloudError as e:
         raise CLIError(unpack_msrest_error(e))
-
-    if add_children:
-        for child_device_id in add_children.split(","):
-            child_device = _iot_device_show(target, child_device_id.strip())
-            _update_device_parent(target, child_device, child_device["capabilities"]["iotEdge"], output.device_scope)
 
     return output
 
@@ -297,7 +302,7 @@ def update_iot_device_custom(
     if status_reason is not None:
         instance["statusReason"] = status_reason
 
-    auth_type = instance['authentication']['type']
+    auth_type = instance["authentication"]["type"]
     if auth_method is not None:
         if auth_method == DeviceAuthType.shared_private_key.name:
             auth = "sas"
@@ -359,24 +364,34 @@ def update_iot_device_custom(
 
 
 def iot_device_update(
-    cmd, device_id, parameters, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    device_id,
+    parameters,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
 
     auth, pk, sk = _parse_auth(parameters)
     updated_device = _assemble_device(
         True,
-        parameters['deviceId'],
+        parameters["deviceId"],
         auth,
-        parameters['capabilities']['iotEdge'],
+        parameters["capabilities"]["iotEdge"],
         pk,
         sk,
-        parameters['status'].lower(),
-        parameters.get('statusReason'),
-        parameters.get('deviceScope')
+        parameters["status"].lower(),
+        parameters.get("statusReason"),
+        parameters.get("deviceScope"),
     )
     updated_device.etag = etag if etag else "*"
     return _iot_device_update(target, device_id, updated_device)
@@ -390,20 +405,27 @@ def _iot_device_update(target, device_id, device):
         headers = {}
         headers["If-Match"] = '"{}"'.format(device.etag)
         return service_sdk.devices.create_or_update_identity(
-            id=device_id,
-            device=device,
-            custom_headers=headers
+            id=device_id, device=device, custom_headers=headers
         )
     except CloudError as e:
         raise CLIError(unpack_msrest_error(e))
 
 
 def iot_device_delete(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -411,9 +433,7 @@ def iot_device_delete(
     try:
         headers = {}
         headers["If-Match"] = '"{}"'.format(etag if etag else "*")
-        service_sdk.devices.delete_identity(
-            id=device_id, custom_headers=headers
-        )
+        service_sdk.devices.delete_identity(id=device_id, custom_headers=headers)
         return
     except CloudError as e:
         raise CLIError(unpack_msrest_error(e))
@@ -437,13 +457,25 @@ def _update_device_key(target, device, auth_method, pk, sk, etag=None):
         raise CLIError(unpack_msrest_error(e))
 
 
-def iot_device_key_regenerate(cmd, hub_name, device_id, renew_key_type, resource_group_name=None, login=None, etag=None):
+def iot_device_key_regenerate(
+    cmd,
+    hub_name,
+    device_id,
+    renew_key_type,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
+):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     device = _iot_device_show(target, device_id)
-    if (device["authentication"]["type"] != "sas"):
+    if device["authentication"]["type"] != "sas":
         raise CLIError("Device authentication should be of type sas")
 
     pk = device["authentication"]["symmetricKey"]["primaryKey"]
@@ -458,15 +490,25 @@ def iot_device_key_regenerate(cmd, hub_name, device_id, renew_key_type, resource
         pk = sk
         sk = temp
 
-    return _update_device_key(target, device, device["authentication"]["type"], pk, sk, etag)
+    return _update_device_key(
+        target, device, device["authentication"]["type"], pk, sk, etag
+    )
 
 
 def iot_device_get_parent(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     child_device = _iot_device_show(target, device_id)
     _validate_child_device(child_device)
@@ -485,17 +527,26 @@ def iot_device_set_parent(
     hub_name=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     parent_device = _iot_device_show(target, parent_id)
     _validate_edge_device(parent_device)
     child_device = _iot_device_show(target, device_id)
     _validate_parent_child_relation(child_device, force)
 
-    _update_device_parent(target, child_device, child_device["capabilities"]["iotEdge"], parent_device["deviceScope"])
+    _update_device_parent(
+        target,
+        child_device,
+        child_device["capabilities"]["iotEdge"],
+        parent_device["deviceScope"],
+    )
 
 
 def iot_device_children_add(
@@ -506,26 +557,31 @@ def iot_device_children_add(
     hub_name=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     devices = []
     edge_device = _iot_device_show(target, device_id)
     _validate_edge_device(edge_device)
     converted_child_list = child_list
-    if isinstance(child_list, str):  # this check would be removed once add-children command is deprecated
-        converted_child_list = child_list.split(",")
     for child_device_id in converted_child_list:
         child_device = _iot_device_show(target, child_device_id.strip())
-        _validate_parent_child_relation(
-            child_device, force
-        )
+        _validate_parent_child_relation(child_device, force)
         devices.append(child_device)
 
     for device in devices:
-        _update_device_parent(target, device, device["capabilities"]["iotEdge"], edge_device["deviceScope"])
+        _update_device_parent(
+            target,
+            device,
+            device["capabilities"]["iotEdge"],
+            edge_device["deviceScope"],
+        )
 
 
 def iot_device_children_remove(
@@ -536,10 +592,14 @@ def iot_device_children_remove(
     hub_name=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     devices = []
     if remove_all:
@@ -559,8 +619,6 @@ def iot_device_children_remove(
         edge_device = _iot_device_show(target, device_id)
         _validate_edge_device(edge_device)
         converted_child_list = child_list
-        if isinstance(child_list, str):  # this check would be removed once remove-children command is deprecated
-            converted_child_list = child_list.split(",")
         for child_device_id in converted_child_list:
             child_device = _iot_device_show(target, child_device_id.strip())
             _validate_child_device(child_device)
@@ -582,42 +640,58 @@ def iot_device_children_remove(
 
 
 def iot_device_children_list(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     result = _iot_device_children_list(
-        cmd, device_id, hub_name, resource_group_name, login
+        cmd=cmd,
+        device_id=device_id,
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
 
     return [device["deviceId"] for device in result]
 
 
-# this method would be removed once remove-children command is deprecated
-def iot_device_children_list_comma_separated(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
-):
-    result = _iot_device_children_list(
-        cmd, device_id, hub_name, resource_group_name, login
-    )
-    if not result:
-        raise CLIError(
-            'No registered child devices found for "{}" edge device.'.format(device_id)
-        )
-    return ", ".join([str(x["deviceId"]) for x in result])
-
-
 def _iot_device_children_list(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     device = _iot_device_show(target, device_id)
     _validate_edge_device(device)
-    query = "select deviceId from devices where array_contains(parentScopes, '{}')".format(
-        device["deviceScope"]
+    query = (
+        "select deviceId from devices where array_contains(parentScopes, '{}')".format(
+            device["deviceScope"]
+        )
     )
-    return iot_query(cmd, query, hub_name, None, resource_group_name, login=login)
+
+    # TODO: Inefficient
+    return iot_query(
+        cmd=cmd,
+        query_command=query,
+        hub_name=hub_name,
+        top=None,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type_dataplane=auth_type_dataplane,
+    )
 
 
 def _update_device_parent(target, device, is_edge, device_scope=None):
@@ -667,9 +741,7 @@ def _validate_child_device(device):
         )
     if not device["parentScopes"]:
         raise CLIError(
-            'Device "{}" doesn\'t have any parent device.'.format(
-                device["deviceId"]
-            )
+            'Device "{}" doesn\'t have any parent device.'.format(device["deviceId"])
         )
 
 
@@ -700,6 +772,7 @@ def iot_device_module_create(
     output_dir=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
 
     if any([valid_days, output_dir]):
@@ -715,7 +788,10 @@ def iot_device_module_create(
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -752,10 +828,14 @@ def iot_device_module_update(
     resource_group_name=None,
     login=None,
     etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -805,11 +885,20 @@ def _parse_auth(parameters):
 
 
 def iot_device_module_list(
-    cmd, device_id, hub_name=None, top=1000, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    top=1000,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -821,11 +910,20 @@ def iot_device_module_list(
 
 
 def iot_device_module_show(
-    cmd, device_id, module_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    module_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     return _iot_device_module_show(target, device_id, module_id)
 
@@ -845,11 +943,21 @@ def _iot_device_module_show(target, device_id, module_id):
 
 
 def iot_device_module_delete(
-    cmd, device_id, module_id, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    device_id,
+    module_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -866,11 +974,20 @@ def iot_device_module_delete(
 
 
 def iot_device_module_twin_show(
-    cmd, device_id, module_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    module_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     return _iot_device_module_twin_show(
         target=target, device_id=device_id, module_id=module_id
@@ -897,13 +1014,17 @@ def iot_device_module_twin_update(
     hub_name=None,
     resource_group_name=None,
     login=None,
-    etag=None
+    etag=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.common.utility import verify_transform
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -938,11 +1059,15 @@ def iot_device_module_twin_replace(
     hub_name=None,
     resource_group_name=None,
     login=None,
-    etag=None
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -962,13 +1087,22 @@ def iot_device_module_twin_replace(
 
 
 def iot_edge_set_modules(
-    cmd, device_id, content, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    content,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.sdk.iothub.service.models import ConfigurationContent
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -999,6 +1133,7 @@ def iot_edge_deployment_create(
     no_validation=False,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     # Short-term fix for --no-validation
     config_type = ConfigType.layered if layered or no_validation else ConfigType.edge
@@ -1014,6 +1149,7 @@ def iot_edge_deployment_create(
         resource_group_name=resource_group_name,
         login=login,
         config_type=config_type,
+        auth_type_dataplane=auth_type_dataplane,
     )
 
 
@@ -1028,6 +1164,7 @@ def iot_hub_configuration_create(
     metrics=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     return _iot_hub_configuration_create(
         cmd=cmd,
@@ -1041,6 +1178,7 @@ def iot_hub_configuration_create(
         resource_group_name=resource_group_name,
         login=login,
         config_type=ConfigType.adm,
+        auth_type_dataplane=auth_type_dataplane,
     )
 
 
@@ -1056,6 +1194,7 @@ def _iot_hub_configuration_create(
     metrics=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.sdk.iothub.service.models import (
         Configuration,
@@ -1065,7 +1204,10 @@ def _iot_hub_configuration_create(
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1206,14 +1348,24 @@ def _validate_payload_schema(content):
 
 
 def iot_hub_configuration_update(
-    cmd, config_id, parameters, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    config_id,
+    parameters,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.sdk.iothub.service.models import Configuration
     from azext_iot.common.utility import verify_transform
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1232,7 +1384,7 @@ def iot_hub_configuration_update(
             content=parameters["content"],
             metrics=parameters.get("metrics", None),
             target_condition=parameters["targetCondition"],
-            priority=parameters["priority"]
+            priority=parameters["priority"],
         )
         return service_sdk.configuration.create_or_update(
             id=config_id, configuration=config, custom_headers=headers
@@ -1244,11 +1396,19 @@ def iot_hub_configuration_update(
 
 
 def iot_hub_configuration_show(
-    cmd, config_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    config_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     return _iot_hub_configuration_show(target=target, config_id=config_id)
 
@@ -1264,13 +1424,19 @@ def _iot_hub_configuration_show(target, config_id):
 
 
 def iot_hub_configuration_list(
-    cmd, hub_name=None, top=None, resource_group_name=None, login=None
+    cmd,
+    hub_name=None,
+    top=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     result = _iot_hub_configuration_list(
-        cmd,
+        cmd=cmd,
         hub_name=hub_name,
         resource_group_name=resource_group_name,
         login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
     filtered = [
         c
@@ -1284,13 +1450,19 @@ def iot_hub_configuration_list(
 
 
 def iot_edge_deployment_list(
-    cmd, hub_name=None, top=None, resource_group_name=None, login=None
+    cmd,
+    hub_name=None,
+    top=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     result = _iot_hub_configuration_list(
         cmd,
         hub_name=hub_name,
         resource_group_name=resource_group_name,
         login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
 
     filtered = [c for c in result if c["content"].get("modulesContent") is not None]
@@ -1298,11 +1470,14 @@ def iot_edge_deployment_list(
 
 
 def _iot_hub_configuration_list(
-    cmd, hub_name=None, resource_group_name=None, login=None
+    cmd, hub_name=None, resource_group_name=None, login=None, auth_type_dataplane=None
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1317,11 +1492,20 @@ def _iot_hub_configuration_list(
 
 
 def iot_hub_configuration_delete(
-    cmd, config_id, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    config_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1342,6 +1526,7 @@ def iot_edge_deployment_metric_show(
     hub_name=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     return iot_hub_configuration_metric_show(
         cmd,
@@ -1351,6 +1536,7 @@ def iot_edge_deployment_metric_show(
         hub_name=hub_name,
         resource_group_name=resource_group_name,
         login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
 
 
@@ -1362,10 +1548,14 @@ def iot_hub_configuration_metric_show(
     hub_name=None,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1407,11 +1597,19 @@ def iot_hub_configuration_metric_show(
 
 
 def iot_device_twin_show(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     return _iot_device_twin_show(target=target, device_id=device_id)
 
@@ -1441,13 +1639,23 @@ def iot_twin_update_custom(instance, desired=None, tags=None):
 
 
 def iot_device_twin_update(
-    cmd, device_id, parameters, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    device_id,
+    parameters,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.common.utility import verify_transform
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1472,11 +1680,21 @@ def iot_device_twin_update(
 
 
 def iot_device_twin_replace(
-    cmd, device_id, target_json, hub_name=None, resource_group_name=None, login=None, etag=None
+    cmd,
+    device_id,
+    target_json,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
+    etag=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1501,6 +1719,7 @@ def iot_device_method(
     timeout=30,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.constants import (
         METHOD_INVOKE_MAX_TIMEOUT_SEC,
@@ -1518,7 +1737,10 @@ def iot_device_method(
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1539,7 +1761,9 @@ def iot_device_method(
         }
 
         return service_sdk.devices.invoke_method(
-            device_id=device_id, direct_method_request=request_body, timeout=timeout,
+            device_id=device_id,
+            direct_method_request=request_body,
+            timeout=timeout,
         )
     except CloudError as e:
         raise CLIError(unpack_msrest_error(e))
@@ -1558,6 +1782,7 @@ def iot_device_module_method(
     timeout=30,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.constants import (
         METHOD_INVOKE_MAX_TIMEOUT_SEC,
@@ -1575,7 +1800,10 @@ def iot_device_module_method(
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        login=login,
+        auth_type=auth_type_dataplane,
     )
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
@@ -1618,6 +1846,7 @@ def iot_get_sas_token(
     resource_group_name=None,
     login=None,
     module_id=None,
+    auth_type_dataplane=None,
 ):
     key_type = key_type.lower()
     policy_name = policy_name.lower()
@@ -1646,6 +1875,7 @@ def iot_get_sas_token(
             duration,
             resource_group_name,
             login,
+            auth_type_dataplane,
         ).generate_sas_token()
     }
 
@@ -1660,11 +1890,16 @@ def _iot_build_sas_token(
     duration=3600,
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     from azext_iot.common._azure import (
         parse_iot_device_connection_string,
         parse_iot_device_module_connection_string,
     )
+
+    # There is no dataplane operation for a pure IoT Hub sas token
+    if all([device_id is None, module_id is None]):
+        auth_type_dataplane = "key"
 
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
@@ -1672,6 +1907,7 @@ def _iot_build_sas_token(
         resource_group_name=resource_group_name,
         policy_name=policy_name,
         login=login,
+        auth_type=auth_type_dataplane,
     )
     uri = None
     policy = None
@@ -1753,6 +1989,7 @@ def iot_get_device_connection_string(
     key_type="primary",
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     result = {}
     device = iot_device_show(
@@ -1761,6 +1998,7 @@ def iot_get_device_connection_string(
         hub_name=hub_name,
         resource_group_name=resource_group_name,
         login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
     result["connectionString"] = _build_device_or_module_connection_string(
         device, key_type
@@ -1776,6 +2014,7 @@ def iot_get_module_connection_string(
     key_type="primary",
     resource_group_name=None,
     login=None,
+    auth_type_dataplane=None,
 ):
     result = {}
     module = iot_device_module_show(
@@ -1785,6 +2024,7 @@ def iot_get_module_connection_string(
         resource_group_name=resource_group_name,
         hub_name=hub_name,
         login=login,
+        auth_type_dataplane=auth_type_dataplane,
     )
     result["connectionString"] = _build_device_or_module_connection_string(
         module, key_type
@@ -2016,18 +2256,24 @@ def _iot_c2d_message_receive(target, device_id, lock_timeout=60, ack=None):
                     ack_response = {}
                     if ack == SettleType.abandon.value:
                         logger.debug("__Abandoning message__")
-                        ack_response = device_sdk.device.abandon_device_bound_notification(
-                            id=device_id, etag=eTag, raw=True
+                        ack_response = (
+                            device_sdk.device.abandon_device_bound_notification(
+                                id=device_id, etag=eTag, raw=True
+                            )
                         )
                     elif ack == SettleType.reject.value:
                         logger.debug("__Rejecting message__")
-                        ack_response = device_sdk.device.complete_device_bound_notification(
-                            id=device_id, etag=eTag, reject="", raw=True
+                        ack_response = (
+                            device_sdk.device.complete_device_bound_notification(
+                                id=device_id, etag=eTag, reject="", raw=True
+                            )
                         )
                     else:
                         logger.debug("__Completing message__")
-                        ack_response = device_sdk.device.complete_device_bound_notification(
-                            id=device_id, etag=eTag, raw=True
+                        ack_response = (
+                            device_sdk.device.complete_device_bound_notification(
+                                id=device_id, etag=eTag, raw=True
+                            )
                         )
 
                     payload["ack"] = (
@@ -2238,7 +2484,11 @@ def iot_simulate_device(
 
 
 def iot_c2d_message_purge(
-    cmd, device_id, hub_name=None, resource_group_name=None, login=None,
+    cmd,
+    device_id,
+    hub_name=None,
+    resource_group_name=None,
+    login=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
@@ -2247,7 +2497,9 @@ def iot_c2d_message_purge(
     resolver = SdkResolver(target=target)
     service_sdk = resolver.get_sdk(SdkType.service_sdk)
 
-    return service_sdk.cloud_to_device_messages.purge_cloud_to_device_message_queue(device_id)
+    return service_sdk.cloud_to_device_messages.purge_cloud_to_device_message_queue(
+        device_id
+    )
 
 
 def _iot_simulate_get_default_properties(protocol):
@@ -2313,7 +2565,9 @@ def iot_device_export(
             authentication_type=storage_authentication_type,
         )
         return client.export_devices(
-            target["resourcegroup"], hub_name, export_devices_parameters=export_request,
+            target["resourcegroup"],
+            hub_name,
+            export_devices_parameters=export_request,
         )
     if storage_authentication_type:
         raise CLIError(
@@ -2366,7 +2620,9 @@ def iot_device_import(
             authentication_type=storage_authentication_type,
         )
         return client.import_devices(
-            target["resourcegroup"], hub_name, import_devices_parameters=import_request,
+            target["resourcegroup"],
+            hub_name,
+            import_devices_parameters=import_request,
         )
     if storage_authentication_type:
         raise CLIError(
@@ -2501,11 +2757,17 @@ def iot_hub_monitor_feedback(
 
 
 def iot_hub_distributed_tracing_show(
-    cmd, hub_name, device_id, resource_group_name=None, login=None,
+    cmd,
+    hub_name,
+    device_id,
+    resource_group_name=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
-        hub_name=hub_name, resource_group_name=resource_group_name, login=login
+        hub_name=hub_name,
+        resource_group_name=resource_group_name,
+        auth_type=auth_type_dataplane,
     )
 
     device_twin = _iot_hub_distributed_tracing_show(target=target, device_id=device_id)
@@ -2598,14 +2860,14 @@ def iot_hub_distributed_tracing_update(
     sampling_mode,
     sampling_rate,
     resource_group_name=None,
-    login=None,
+    auth_type_dataplane=None,
 ):
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
         hub_name=hub_name,
         resource_group_name=resource_group_name,
         include_events=True,
-        login=login,
+        auth_type=auth_type_dataplane,
     )
 
     if int(sampling_rate) not in range(0, 101):
@@ -2622,7 +2884,7 @@ def iot_hub_distributed_tracing_update(
         1 if sampling_mode.lower() == "on" else 2
     )
     result = iot_device_twin_update(
-        cmd, device_id, device_twin, hub_name, resource_group_name, login
+        cmd, device_id, device_twin, hub_name, resource_group_name
     )
     return _customize_device_tracing_output(
         result.device_id, result.properties.desired, result.properties.reported
@@ -2654,20 +2916,26 @@ def iot_hub_connection_string_show(
         for hub in hubs:
             if hub.properties.state == IoTHubStateType.Active.value:
                 try:
-                    connection_strings.append({
-                        "name": hub.name,
-                        "connectionString": conn_str_getter(hub)
-                        if show_all
-                        else conn_str_getter(hub)[0],
-                    })
+                    connection_strings.append(
+                        {
+                            "name": hub.name,
+                            "connectionString": conn_str_getter(hub)
+                            if show_all
+                            else conn_str_getter(hub)[0],
+                        }
+                    )
                 except:
-                    logger.warning(f"Warning: The IoT Hub {hub.name} in resource group " +
-                                   f"{hub.additional_properties['resourcegroup']} does " +
-                                   f"not have the target policy {policy_name}.")
+                    logger.warning(
+                        f"Warning: The IoT Hub {hub.name} in resource group "
+                        + f"{hub.additional_properties['resourcegroup']} does "
+                        + f"not have the target policy {policy_name}."
+                    )
             else:
-                logger.warning(f"Warning: The IoT Hub {hub.name} in resource group " +
-                               f"{hub.additional_properties['resourcegroup']} is skipped " +
-                               "because the hub is not active.")
+                logger.warning(
+                    f"Warning: The IoT Hub {hub.name} in resource group "
+                    + f"{hub.additional_properties['resourcegroup']} is skipped "
+                    + "because the hub is not active."
+                )
         return connection_strings
 
     hub = discovery.find_iothub(hub_name, resource_group_name)
@@ -2710,7 +2978,12 @@ def _get_hub_connection_string(
                 entityPath,
             )
             for p in policies
-            if "serviceconnect" in (p.rights.value.lower() if isinstance(p.rights, (Enum, EnumMeta)) else p.rights.lower())
+            if "serviceconnect"
+            in (
+                p.rights.value.lower()
+                if isinstance(p.rights, (Enum, EnumMeta))
+                else p.rights.lower()
+            )
         ]
 
     hostname = hub.properties.host_name